+++ src/Resources/Policy.Autorest/test/Common.ps1
@@ -99,6 +99,7 @@ if ($TargetTestName -and (!$TestName -or ($TestName -eq $TargetTestName))) {
         $policyAssignmentMissingIdentityId = $env.policyAssignmentMissingIdentityId
         $policyExemptionNotFound = $env.policyExemptionNotFound
         $policyEnrollmentNotFound = $env.policyEnrollmentNotFound
+        $unableToUpdateEnrollment = $env.unableToUpdateEnrollment
         $invalidRequestContent = $env.invalidRequestContent
         $policyDefinitionParameter = $env.policyDefinitionParameter
         $missingSubscription = $env.missingSubscription